adds methods for non GC mode support; GLMesh no longer always requests new buffers; adds some primitive integer vector operations

+++ source/SocketMessenger.m
@@ -393,7 +393,7 @@ SEND_ERR:
 	NSInteger	port = [remoteService port];
 	NSString*	hostName = [remoteService hostName];
 	
-	NSLog(@"SocketMessenger attempting to connect to %@ : %d", hostName, port);
+	NSLog(@"SocketMessenger attempting to connect to %@ : %d", hostName, (int)port);
 
 	int sockfd = 0;
